Updated GkTypesLib Dependency

+++ Dependencies/include/GkTypes/Hashmap/Hash.h
@@ -5,16 +5,16 @@
 namespace gk 
 {
 	template<typename KeyType>
-	constexpr size_t hash(KeyType key) = delete;
+	constexpr uint64 hash(KeyType key) = delete;
 
 	template<>
-	constexpr size_t hash<size_t>(size_t key) {
+	constexpr uint64 hash<size_t>(size_t key) {
 		return key;
 	}
 
 	template<>
-	constexpr size_t hash<int>(int key) {
-		return size_t(key);
+	constexpr uint64 hash<int>(int key) {
+		return uint64(key);
 	}
 
 }
